fixed bugs on WeatherDataDto.kt it now runs

+++ app/src/main/java/com/plcoding/weatherapp/data/remote/WeatherDataDto.kt
@@ -7,7 +7,7 @@ data class WeatherDataDto(
     @field:Json(name = "temperature_2m")
     val temperatures: List<Double>,
     @field:Json(name = "weathercode")
-    val weatherCodes: List<Double>,
+    val weatherCodes: List<Int>,
     @field:Json(name = "pressure_msl")
     val pressure: List<Double>,
     @field:Json(name = "windspeed_10m")
